+++ activerecord/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Schema loading rake tasks (like `db:schema:load` and `db:setup`) maintain
+    the database connection to the current environment.
+
+    Fixes #16757.
+
+    *Joshua Cody*, *Yves Senn*
+
 *   `db:purge` with MySQL respects `Rails.env`.
 
     *Yves Senn*